Update patch set 4
Patch Set 4:

Thanks for the work Ralph, here are some comments, my intent is that they are constructive:

The code is well documented, and there is the clear intent to performing operations asynchronously.  The resulting drl simplifies to a large degree what we have over our current base, and I can see how the approach will improve performance.  

To keep things in perspective, the final goal of this contribution is that it is usable in the ONAP use cases, such as the scale out, vFirewall, etc .. and that it significantly improves our code base.

On the pending work side, the review does not integrate just yet with the existing ONAP base.   I understand that the proposal is that this will be done later on after this submission, and from my side I would be OK with it.  The concern is that it seems to me there is quite a bit of work left.

In any case, these are the functional areas for what I see in this review that still need to be tackled:

1. Integration with usecases style drl with ONAP flows.  In this drl, there is not template expansion, no need to store loggers in contexts, etc.  

2. Simulators, need to reconcile existing and new provided implementations, determine which one is better, and use only 1 to avoid the code duplication.

3. Guard, I think there are still 2 versions of the code updating the DB.

4. I think similarly for the Appc actor, it is my impression that there is still some overlap with that in policy-models, although I am not entirely sure.

From my side, I would be OK staging the pending work after this one, as long as there is a firm commitment to get it accomplished.   This agreement could be documented in the form of high priority JIRA tickets and would need to be tackled right after this review in Frankfurt.

Another request is to encapsulate this work on its own feature, so:

5. Encapsulate all the work on an m2 feature, to make this functionality toggable and therefore optional.

In addition, since these projects are "control loop" related, they should be moved under the controlloop directory.

With regards to 5 and 6., similarly, I am OK with postponing it to right after this review.

There is also the sonar and junit work to keep up with the numbers currently have in this repo.

As an aside, something that worries me somewhat is the parallel work on the actor clean up implementation task that may be going on in parallel, that would also need to be synced at some point.

Patch-set: 4
Label: Code-Review=0