Update patch set 4
Patch Set 4:

(1 comment)

> Patch Set 4:
> 
> Thanks for the work Ralph, here are some comments, my intent is that they are constructive:
> 
> The code is well documented, and there is the clear intent to performing operations asynchronously.  The resulting drl simplifies to a large degree what we have over our current base, and I can see how the approach will improve performance.  
> 
> To keep things in perspective, the final goal of this contribution is that it is usable in the ONAP use cases, such as the scale out, vFirewall, etc .. and that it significantly improves our code base.
> 
> On the pending work side, the review does not integrate just yet with the existing ONAP base.   I understand that the proposal is that this will be done later on after this submission, and from my side I would be OK with it.  The concern is that it seems to me there is quite a bit of work left.
> 
> In any case, these are the functional areas for what I see in this review that still need to be tackled:
> 
> 1. Integration with usecases style drl with ONAP flows.  In this drl, there is not template expansion, no need to store loggers in contexts, etc.  
> 
> 2. Simulators, need to reconcile existing and new provided implementations, determine which one is better, and use only 1 to avoid the code duplication.
> 
> 3. Guard, I think there are still 2 versions of the code updating the DB.
> 
> 4. I think similarly for the Appc actor, it is my impression that there is still some overlap with that in policy-models, although I am not entirely sure.
> 
> From my side, I would be OK staging the pending work after this one, as long as there is a firm commitment to get it accomplished.   This agreement could be documented in the form of high priority JIRA tickets and would need to be tackled right after this review in Frankfurt.
> 
> Another request is to encapsulate this work on its own feature, so:
> 
> 5. Encapsulate all the work on an m2 feature, to make this functionality toggable and therefore optional.
> 
> In addition, since these projects are "control loop" related, they should be moved under the controlloop directory.
> 
> With regards to 5 and 6., similarly, I am OK with postponing it to right after this review.
> 
> There is also the sonar and junit work to keep up with the numbers currently have in this repo.
> 
> As an aside, something that worries me somewhat is the parallel work on the actor clean up implementation task that may be going on in parallel, that would also need to be synced at some point.

Regarding integration with ONAP -- this is a new high-level framework (new to ONAP -- it is actually about 2 1/2 years old). It had the following design goals (taken from the m2 model viewgraph presentation):

1) Use a single object in Drools memory for each transaction

2) Make the state more explicit

3) Incoming events are transient

4) No actor-specific or operator-specific code within the transaction object

Much of the low-level code was reused, but the high-level control code in the older model was completely replaced. Everything was designed around the 'Transaction', 'Actor', and 'Operation' classes, and these were designed with performance and extensibility in mind.

Regarding integration with ONAP, the big question is whether there is general buy-in to this architecture. If people agree that this is the way to go, "integration with ONAP" involves adding other 'Actor' and 'Operation' subclasses to cover current ONAP functionality that does not yet exist in m2. If people don't agree, then this was a wasted effort, and it would be better to not merge this change.

Regarding item #1, the only '.drl' file we have currently provided is one used in the 'LcmTest' Junit test. We do have other ECOMP-specific files. All of them are currently based upon templates, but we do plan to move away from that.

Regarding item #2, I will try to find time to look at and try out existing simulators. However, the ONAP simulators seem to have an order of magnitude more code, which I don't regard as a good sign.

Regarding item #3, I did look at the existing code updating the DB, and it is in 'ControlLoopOperationManager', deep within the bowels of the existing model. I didn't see any good way to call this code, but I did make use of the 'Dbao' class. The query portion was easier to reuse, because it wasn't coupled to the existing model.

Regarding item #4, we are using ONAP lower-level code, but the higher-level control code needs to be compatible with the m2 control structure, so it is different. 'LcmResponseCode' might be the one class that could be shared -- I don't remember the details, but I probably looked at this earlier, and decided we needed a different class.

Regarding item #5, we did discuss this earlier, and I agree that we should move the code and make it a feature. I was hoping to do this after we resolved other issues, because moving code makes it more difficult to compare patch sets.

Patch-set: 4
Label: Code-Review=0