Update patch set 10
Patch Set 10:

> Patch Set 10: -Code-Review
> 
> How is the code coverage on all this work?

Good, ~50% at the moment with the submission.  We have several developers on the team working and ready to add the remaining junits asap to get to 80% once this is submitted.  We divided the work for junits to get it done quicker.  So can we get this merged and finish this off this week?  Thanks much.

Patch-set: 10