Update patch set 4
Patch Set 4:

Peyton, were discussing this error with committers in chat, this one that you're picking up some errors, as powermock-api-mockito is not included.   I think this may have to do with move to jdk 11.   In any case, I guess we need you to use powermock-api-mockito2 dependency instead which is the one in policy/parent.

Patch-set: 4
Label: Verified=0