Update patch set 4
Patch Set 4:

> (1 comment)
 > 
 > Should we be pulling example policies from models-examples? Instead
 > of having multiple copies in separate repos?

It is on my list to do, I am give higher priority though on getting the core functionality done, and then circle back to this.   I recall that there were some differences on some of them though, and the policies are embedded into these service templates.   So, the answer is yes, but later, and may require changes in policy/models too if the policy data does not match the expectations of the simulators.

Patch-set: 4
Label: Code-Review=0