Update patch set 2
Patch Set 2:

Need some changes in the Dockerfile to remove the amsterdam lines, and can add a line to add "frankfurt" one that I missed to comment in Jim's review.

I can do some changes in OOM to remove any amsterdam references.

Do a global search for amsterdam in the repo, I think there are some scripts such as "create-cl-amsterdam" that will need to be removed too.

But let's make sure we are all in agreement with my previous questions.

Patch-set: 2
Label: Verified=0