Fix db exception in frankfurt junit
Frankfurt junits were generating the following output:

  Value too long for column "TARGET VARCHAR(50)":
    "'Target [type=VNF, resourceId=bbb3cefd-01c8-413c-9bdd-2b92f9ca3d38]'

The issue turned out to be caused by the fact that the rules for the
new actor were storing the "Target" instead of the "TargetEntity"
in the DB field.  Modified the code to store the target entity, as the
the usecases rules do.

Issue-ID: POLICY-2441
Signed-off-by: Jim Hahn <jrh3@att.com>
Change-Id: I91ee517ef073e3dc3fea4698c814b57a06d87095