Handle duplicate events in drools-apps
Modified the frankfurt rules to compare the event objects instead of
comparing the request ID when determining if an event is new.  The
event object's equals() method ignores the request ID when doing the
comparison, thus it will treat an event as a duplicate even if the
request ID is different, which is the behavior we want.

Also removed the @Ignore from the junit that tests for duplicate
events in the hope that this change will fix it.  If the docker build
still breaks, then @Ignore can be added back in.

Issue-ID: POLICY-2557
Change-Id: If2b9fd26473d78a356218b951bfe160f93daeb32
Signed-off-by: Jim Hahn <jrh3@att.com>