Use "outcome" to indicate pending record
Mariadb is setting the "endtime" to the current time by default,
so rather than indicate a pending record using a null endtime, just
modified the code to use null "outcome" to indicate a pending record.
Also addressed a comment from a previous review to log a warning message
if more than one matching record is found in operations history.

Issue-ID: POLICY-2581
Change-Id: I6bf67551cef46808a79cc15afeb0abbfc1b6d945
Signed-off-by: Jim Hahn <jrh3@att.com>