Do additional processing when control loop times out
When the overall control loop times out, a notification is sent to
POLICY-CL-MGT, but then the event is discarded.  Modified the code
to update the operation history, if an operation had been started,
and to generate an SDNR notification, if appropriate.
Applied the same logic for when a lock is lost.

Issue-ID: POLICY-2582
Change-Id: I35510ec8a3cfc6b2ca9be5751896f3c0f7d9bd09
Signed-off-by: Jim Hahn <jrh3@att.com>