Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1: Code-Review+1
> 
> Yuck - I don't really like hardcoding actors into the ControlLoopEventManager. But I guess we have no choice right now.

My original take on things was similar to yours, which is why there is currently code, within the SDNR actor, to handle some of the SDNR notifications.  However, the more I think about it, it seems to me that notifications are really more of an application/company-specific thing, which would imply that they should be kept in the application code rather than the actor code.  For that matter, the fact that it's only done for SDNR makes me more inclined toward handling this in the application code.
So later in Guilin, I'm thinking of moving the existing notification code out of the SDNR actor and consolidating it within the application.  However, if you think this is the wrong approach, I can consolidate it within the actor, instead.

Patch-set: 1
Reviewer: Gerrit User 811 <811@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>