Update patch set 2
Patch Set 2: Code-Review-2

Too soon to rename frankfurt controller - we do want to be able to do a release with all the tech debt leftover from frankfurt. Not sure if this is premature here to do this.

What is the plan to get rid of m2? That should be worked on immediately.

Patch-set: 2
Reviewer: Gerrit User 57 <57@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-2