Add tdjam-controller
the details is on wiki:

https://wiki.onap.org/display/DW/tdjam+Feature+for+Tosca-Driven+Control+Loops

Fixed some eclipse warnings.
Revised the interaction between a manager and its SerialWorkQueue to
address some threading issues.  The original code started processing
the event as soon as the manager was created, WHILE it was still being
added to the map.  During junit tests, the event responses came back,
within the same thread, which then attempted to remove the manager from
the map.  This resulted in a ConcurrentHashMap exception.

Issue-ID: POLICY-2415
Change-Id: I94a4152637be76e5b2aea2d869afd84dfb413a0e
Signed-off-by: Straubs, Ralph (rs8887) <rs8887@att.com>
Signed-off-by: jhh <jorge.hernandez-herrero@att.com>
Signed-off-by: Taka Cho <takamune.cho@att.com>
Signed-off-by: Jim Hahn <jrh3@att.com>