Add tdjam-controller
the details is on wiki:

https://wiki.onap.org/display/DW/tdjam+Feature+for+Tosca-Driven+Control+Loops

Fixed some eclipse warnings.
Revised the interaction between a manager and its SerialWorkQueue to
address some threading issues.  The original code started processing
the event as soon as the manager was created, WHILE it was still being
added to the map.  During junit tests, the event responses came back,
within the same thread, which then attempted to remove the manager from
the map.  This resulted in a ConcurrentHashMap exception.

Issue-ID: POLICY-2415
Change-Id: I4eb9d9b3970d22d3653f57b7dc10ef009abf2a57
Signed-off-by: Jim Hahn <jrh3@att.com>