Add "special" Operation classes for new usecases
Created place-holder for new "usecases" controller.  Note: as of yet,
this is still just a normal java project; it will be turned into a
full-fledged kie-based controller in a later review.
Added some Operation and a few other support classes that will be
needed.
Note: the detmTarget() method and it's various supporting methods,
detmXxx(), found in GetTargetEntityOperation2 were cloned from the
pre-existing ControlLoopOperationManager2 class, and modified to fit the
strategy of the new controller.  Likewise for the junit tests of those
methods.

Issue-ID: POLICY-2748
Change-Id: Idffa135a11a56bd6afdd2728706fd40168e80912
Signed-off-by: Jim Hahn <jrh3@att.com>