Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> There is progress here - but vCPE fails because it doesn't have vserver.vserver-name. Which is kind of making me scratch my head a bit.

Ok so A&AI actor is trying to do the custom query here, which would not be needed for vCPE. So how do we silently ignore it for filters? We can turn off guard for this use case and see if they want to support guard. If so, they would need to provide the vserver-name. I was thinking maybe we just return empty attributes in guard and let things go forward. There was another A&AI call to get vserver information based on vserver-id. But we would have to add that.

Patch-set: 10