Release locks between junit tests
Tdjam junits were randomly failing.  Traced it down to the fact that the
locks were being released asynchronously and thus were not always
released when the next test cases requested them.
Also simplified logging for tdjam junit tests.
Modified code to not make a copy of the tdjam property file so the
version stored in gerrit is left untouched.

Issue-ID: POLICY-2789
Change-Id: I325ec69cf7affa531d3c575e3a34bc0b0e1edac7
Signed-off-by: Jim Hahn <jrh3@att.com>