Fix build error in drools-apps related to services
Apparently, creating the event services in the rule file has broken the
junit test of the rules.  It appears that the http clients are not yet
loaded into the drools-pdp engine by the time the first rule fires, thus
the event manager is unable to initialize all of its actors.

Modified the junit to not start the rules until after the http clients
have been started.

Issue-ID: POLICY-3260
Change-Id: I83fe4ea9fa7ca24396019c18f0d77352fbe86db4
Signed-off-by: Jim Hahn <jrh3@att.com>