Fix drools-applications csit test cases
In the recent runs, it was found that controller might take fews more seconds to come up and get reported by the telemetry api. So, increasing the "max" wait time for the test cases to 1 min. And it won't slow down the tests as the check will be made every 15 sec to verify if the assertions are passing. The moment they pass, it will move on.

Issue-ID: POLICY-3257
Change-Id: If29738cae60a6a21da11fa7a6045b4279012b82e
Signed-off-by: Ram Krishna Verma <ram_krishna.verma@bell.ca>