Update patch set 1
Patch Set 1:

Hi All,

I would like to release this image and update the OOM charts as well since this bug needs to be fixed for J-release. I found that the docker build was failing for weeks now 'https://jenkins.onap.org/job/policy-drools-applications-maven-docker-stage-master'. Can this job be fixed so that I can proceed with the image release? (Not sure if someone is already working on this)

Patch-set: 1